Hibernate integration to h2 database
h2 database is built in the memory app. It is dropped with every restart

+++ to-dos-spring-BACK/src/main/java/com/example/demo/helloworld/toDo/ToDoHardcodedService.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Service;
 public class ToDoHardcodedService {
 	
 	private static List<ToDo> toDos = new ArrayList();
-	private static int idCounter=0;
+	private static Long idCounter=(long) 0;
 	
 	
 	static {
